+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/accountinfo/ReleaseSchedule.java
@@ -1,5 +1,6 @@
 package com.concordium.sdk.responses.accountinfo;
 
+import com.concordium.sdk.transactions.GTUAmount;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.ToString;
@@ -9,10 +10,10 @@ import java.util.List;
 @Getter
 @ToString
 public final class ReleaseSchedule {
-    private final String total;
+    private final GTUAmount total;
     private final List<ScheduledRelease> schedule;
 
-    ReleaseSchedule(@JsonProperty("total") String total,
+    ReleaseSchedule(@JsonProperty("total") GTUAmount total,
                     @JsonProperty("schedule") List<ScheduledRelease> schedule) {
         this.total = total;
         this.schedule = schedule;